\name{one.gene}
\alias{one.gene}
\docType{data}
\title{
Score statistics, covariance matrix, and SNP weights for a gene
}
\description{
The data are derived from the genetic association study for a gene. 
}
\usage{data("one.gene")}
\format{
one.gene contains the following objects:
\itemize{
\item U: a vector of score statistics
\item V: a covariance matrix 
\item weight: a vector of SNP weights
}
}
\details{

}
\source{
Tang, Z.-Z., Sliwoski, G., Chen, G., Jin, B., Bush, W., Li, B., and Capra, T. (2020). Spatial scan tests guided by protein structures improve complex disease gene discovery and signal variant detection. Submitted.

}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(one.gene)
}
\keyword{datasets}
