\name{one.gene}
\alias{one.gene}
\docType{data}
\title{
Score statistics, covariance matrix, and SNP weights for a gene
}
\description{
The data are derived from the genetic association study for a gene. 
}
\usage{data("one.gene")}
\format{
one.gene contains the following objects:
\itemize{
\item U: a vector of score statistics
\item V: a covariance matrix 
\item weight: a vector of SNP weights
}
}
\examples{
data(one.gene)
}
\keyword{datasets}
